refactor: Switch to union type for input definitions
