+++ packages/cli/src/commands/add/plugin/add-plugin.input-definitions.ts
